Implement pipeline registers, etc
Implement basic pipeline registers and set up the pipeline to use
them.

Etc -> Got carried away with things and am in a rush, so also renamed
InstructionRegisters enum to InstructionRegister, switched to
unsigned long types for user input, and also started storing the PC
for use by the ROB later. Maybe some other things I'm missing from
glancing at the diffs right now.

+++ src/main.cpp
@@ -13,8 +13,8 @@ int main(const int argc, const char *argv[])
 		return EXIT_FAILURE;
 	}
 
-	Simulator simulator{std::stoi(argv[1]), std::stoi(argv[2]), std::stoi(argv[3]),
-	                    argv[4]};
+	Simulator simulator{std::stoul(argv[1]), std::stoul(argv[2]),
+	                    std::stoul(argv[3]), argv[4]};
 
 	simulator.setup();
 	// simulator.run();
